chore: Update the bot so that it works with latest version of Stumble Guys
Bot did not exit after the first round loss, it would spectate for the following 2 rounds. The "Leave" button can no longer be clicked, the escape keypress is used to exit after the loss. So, F10 is now used to kill the bot instead of the escape key. The bot also did not handle the Stumble Pass and Stumble Journey rewards, so code has been added to detect those screens and proceed.